+++ admin/meta-box-gateway-config.php
@@ -24,25 +24,8 @@ global $pronamic_pay_providers;
 
 bind_providers_and_gateways();
 
-foreach ( $pronamic_pay_providers as $provider ) {
-	$group = array(
-		'name'    => $provider['name'],
-		'options' => array(),
-	);
-
-	if ( isset( $provider['gateways'] ) ) {
-		foreach ( $provider['gateways'] as $id => $gateway ) {
-			$group['options'][ $id ] = $gateway['name'];
-		}
-	}
-
-	$options[] = $group;
-}
-
 ?>
 <div id="pronamic-pay-gateway-config-editor">
-	<?php wp_nonce_field( 'pronamic_pay_save_gateway', 'pronamic_pay_nonce' ); ?>
-
 	<table class="form-table">
 		<tr>
 			<th scope="row">
@@ -57,14 +40,15 @@ foreach ( $pronamic_pay_providers as $provider ) {
 					<?php
 
 					foreach ( $pronamic_pay_providers as $provider ) {
-						if ( isset( $provider['gateways'] ) && is_array( $provider['gateways'] ) ) {
+						if ( isset( $provider['integrations'] ) && is_array( $provider['integrations'] ) ) {
 							printf( '<optgroup label="%s">', esc_attr( $provider['name'] ) );
 
-							foreach ( $provider['gateways'] as $id => $gateway ) {
+							foreach ( $provider['integrations'] as $integration ) {
+								$id      = $integration->get_id();
+								$name    = $integration->get_name();
 								$classes = array();
-								$name    = $gateway['name'];
 
-								if ( isset( $gateway['deprecated'] ) && $gateway['deprecated'] ) {
+								if ( isset( $integration->deprecated ) && $integration->deprecated  ) {
 									$classes[] = 'deprecated';
 
 									$name = sprintf( __( '%s (obsoleted)', 'pronamic_ideal' ), $name );
@@ -75,8 +59,8 @@ foreach ( $pronamic_pay_providers as $provider ) {
 								}
 
 								printf(
-									'<option data-ideal-method="%s" value="%s" %s class="%s">%s</option>',
-									esc_attr( $gateway['gateway'] ),
+									'<option data-pronamic-pay-settings="%s" value="%s" %s class="%s">%s</option>',
+									esc_attr( wp_json_encode( $integration->get_settings() ) ),
 									esc_attr( $id ),
 									selected( $variant_id, $id, false ),
 									esc_attr( implode( ' ', $classes ) ),
@@ -103,7 +87,7 @@ foreach ( $pronamic_pay_providers as $provider ) {
 			$classes[] = 'extra-settings';
 
 			foreach ( $section['methods'] as $method ) {
-				$classes[] = 'method-' . $method;
+				$classes[] = 'setting-' . $method;
 			}
 		}
 
@@ -139,7 +123,7 @@ foreach ( $pronamic_pay_providers as $provider ) {
 						$classes[] = 'extra-settings';
 
 						foreach ( $field['methods'] as $method ) {
-							$classes[] = 'method-' . $method;
+							$classes[] = 'setting-' . $method;
 						}
 					}