+++ classes/Pronamic/WP/Pay/GatewayIntegrations.php
@@ -32,8 +32,18 @@ class Pronamic_WP_Pay_GatewayIntegrations implements IteratorAggregate {
 			$this->integrations[ $object->get_id() ] = $object;
 
 			Pronamic_WP_Pay_ConfigProvider::register( $object->get_id(), $object->get_config_factory_class() );
+		}
+	}
 
-			Pronamic_WP_Pay_GatewayFactory::register( $object->get_config_class(), $object->get_gateway_class() );
+	/**
+	 * Get an integration by the specified ID.
+	 *
+	 * @param string $id
+	 * @return mixed
+	 */
+	public function get_integration( $id ) {
+		if ( isset( $this->integrations[ $id ] ) ) {
+			return $this->integrations[ $id ];
 		}
 	}